Remove is_contractor field from User model
Previously being a contractor user was determined by is_contractor flag
in the user, and being assigned as a contractor for some contract zone.
Because we don't really need the flag, from now on being a contractor
is determined solely based on whether the user is assigned to a
contract zone or not.